requirements.txt
Cython==0.29.28
faiss==1.7.2
h5py==3.6.0
info_nce_pytorch==0.1.4
matplotlib==3.5.2
numpy==1.22.3
open3d==0.15.2
opencv_python==4.6.0.66
pandas==1.4.2
Pillow==8.2.0
Pillow==9.3.0
Pillow==9.5.0
plyfile==0.7.4
pytorch_lightning==0.8.5
scikit_learn==1.1.1
scipy==1.8.1
setuptools==61.2.0
tensorflow==2.12.0
torch==1.8.1
torchvision==0.9.1+cu102
torchvision==0.2.2
tqdm==4.64.0
